split up message direction and Sidebar tabs
