+++ extensions/void/package.json
@@ -363,6 +363,7 @@
     "@types/node": "^22.5.1",
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
+    "@types/uuid": "^10.0.0",
     "@types/vscode": "1.89.0",
     "@typescript-eslint/eslint-plugin": "^8.3.0",
     "@typescript-eslint/parser": "^8.3.0",
@@ -382,13 +383,14 @@
     "rimraf": "^6.0.1",
     "tailwindcss": "^3.4.10",
     "typescript": "5.5.4",
-    "typescript-eslint": "^8.3.0"
+    "typescript-eslint": "^8.3.0",
+    "uuid": "^10.0.0"
   },
   "dependencies": {
     "@anthropic-ai/sdk": "^0.27.1",
+    "diff": "^7.0.0",
     "diff-match-patch": "^1.0.5",
     "ollama": "^0.5.9",
-    "openai": "^4.57.0",
-    "diff": "^7.0.0"
+    "openai": "^4.57.0"
   }
 }
